Cleaned up
Moved classes to their own files/folders.
Changed name to better describe the test fixture class.
