+++ packages/assemblyscript/assembly/wasa/mock/random/xor.ts
